# Speed up Greenlet creation on CPython